+++ mocks/mocks.go
@@ -24,7 +24,7 @@ func NewMockRoundTripper(responses []http.Response) MockMultiResponseRoundTrippe
 func (mrt *MockMultiResponseRoundTripper) Increment() {
 	mrt.mtx.Lock()
 	defer mrt.mtx.Unlock()
-	mrt.calls += 1
+	mrt.calls++
 }
 
 func (mrt *MockMultiResponseRoundTripper) RoundTrip(req *http.Request) (*http.Response, error) {
@@ -72,4 +72,17 @@ func (erc ErrorReadCloser) Read(p []byte) (n int, err error) {
 
 func (erc ErrorReadCloser) Close() error {
 	return nil
+}
+
+// MockEnvGetter implements EnvGetter for testing
+type MockEnvGetter struct {
+	envVars map[string]string
+}
+
+func NewMockEnvGetter(envVars map[string]string) *MockEnvGetter {
+	return &MockEnvGetter{envVars: envVars}
+}
+
+func (m *MockEnvGetter) Getenv(key string) string {
+	return m.envVars[key]
 }
